447 - calculation of ordering was changed

+++ commons-app/src/main/java/cz/cuni/mff/xrg/odcs/commons/app/JobsTypes.java
@@ -4,8 +4,8 @@ public class JobsTypes {
     // -1 - running
     // 0 - normal
     // 1 - unlimited aka realtime
-    public static final int UNLIMITED = 1;
+    public static final long UNLIMITED = 1;
 
-    public static final int MAX_PRIORITY = 0;
+    public static final long MAX_PRIORITY = 0;
 
 }
